rbd-wnbd: introduce RbdMapping class
We're moving most of the WNBD mapping handling to a separate
class called RbdMapping. This simplifies cleanup and makes it
easier to reuse.

The WnbdHandler class covers WNBD specific operations and IO
callbacks while the RbdMapping wrapper will take care of RBD
operations.

A subsequent change will make use of it while switching from
one process per mapping to a single process per host.

While at it, we're also moving the rbd-wnbd config helpers
to separate files.

(cherry picked from commit 96e8850ff3b6661c86eb361051282c8d66ca2031)
