+++ src/tools/rbd_wnbd/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_executable(
     rbd-wnbd
+    rbd_mapping.cc rbd_mapping_config.cc
     rbd_wnbd.cc wnbd_handler.cc wnbd_wmi.cc
     ../../common/win32/code_page.rc)
 set_target_properties(
